requirements.txt
python-telegram-bot==21.4
aiohttp~=3.10.4
python-dotenv~=1.0.1